Add ike encode/decode UT
